feat: Add dynamic clock to display current date and time
The code changes in `time.js` introduce a dynamic clock feature that displays the current date and time on the webpage. This is achieved by retrieving the current date and time using JavaScript's `Date` object and updating the corresponding HTML elements with the retrieved values.

This commit also includes updates to the `menu.js` and `configs.js` files to update the version number and add greetings